move glhelpers to misc.h
