+++ misc.h
@@ -24,6 +24,7 @@
 #include <syslog.h>
 #include <stdarg.h>
 #include <stdio.h>
+#include <stdlib.h>
 #include <time.h>	// clock_gettime
 #include <unistd.h>
 #include <sys/syscall.h>
@@ -31,6 +32,55 @@
 #include <inttypes.h>
 #include <stdint.h>
 
+extern "C" {
+#include <libavcodec/avcodec.h>
+}
+
+#ifdef USE_GLES
+#include <EGL/egl.h>
+/* Hack:
+ * xlib.h via eglplatform.h: #define Status int
+ * X.h via eglplatform.h: #define CurrentTime 0L
+ *
+ * revert it, because it conflicts with vdr variables.
+ */
+#undef Status
+#undef CurrentTime
+#include <GLES2/gl2.h>
+
+#include "logger.h"
+
+/****************************************************************************************
+ * GL Helpers
+ ***************************************************************************************/
+static inline void glCheckError(const char *stmt, const char *fname, int line) {
+	GLint err = glGetError();
+	if (err != GL_NO_ERROR)
+		LOGERROR("GL Error (0x%08x): %s failed at %s:%i", err, stmt, fname, line);
+}
+
+static inline void eglCheckError(const char *stmt, const char *fname, int line) {
+	EGLint err = eglGetError();
+	if (err != EGL_SUCCESS)
+		LOGERROR("EGL ERROR (0x%08x): %s failed at %s:%i", err, stmt, fname, line);
+}
+
+#define GL_CHECK(stmt) do { \
+		stmt; \
+		glCheckError(#stmt, __FILE__, __LINE__); \
+	} while (0)
+
+#define EGL_CHECK(stmt) do { \
+		stmt; \
+		eglCheckError(#stmt, __FILE__, __LINE__); \
+	} while (0)
+
+#endif // USE_GLES
+
+
+/****************************************************************************************
+ * FFmpeg Helpers
+ ***************************************************************************************/
 #ifndef AV_NOPTS_VALUE
 #define AV_NOPTS_VALUE INT64_C(0x8000000000000000)
 #endif
